--initCommit; Move some file based functions to ManagedFileBasedContainer.
